Adding Challenges
I am adding the pairs challenge and the triplets challenge. The
triplets challenge is not fully solved in that it is completely right,
but does not solve larger input problems in a timely manner. I will be
optimizing over the next week to try and improve the running time (time
permitting). After that I may seek help for optimization as needed.